docs: update virtio-blk documentation
Update the documentation for the virtio-blk device to be more accurate.
The user may include 1 or more disks, where the first is the root disk
and any subsequent disks will be used as data disks. Also do not restrict
the backing image to be a raw image, as we now support QCOW2.

Resolves: #48

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>