build.rs: link with library specified by LIBKRUN_EFI
Rather than hardcoding a path to the /opt/homebrew/lib directory for the
build.rs file to link with, we should use the path provided by the user
in LIBKRUN_EFI instead. If the environment variable cannot be found
because a user didn't use it, then fallback to the default path.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>