Set mac address before gvproxy path
With new libkrun if we set the mac address after setting gvproxy path,
the network device is created using a default mac address:

    let mac = cfg.mac.unwrap_or([0x5a, 0x94, 0xef, 0xe4, 0x0c, 0xee]);
    create_virtio_net(
        cfg,
        VirtioNetBackend::UnixgramPath(path, true),
        mac,
        NET_COMPAT_FEATURES,
    );

This breaks programs like vment-helper tests and minikube, assuming that
the device will use the mac address set by the tool.

This also breaks krunkit networking when using vment-helper with
offloading. I'm not sure why this happens.

With this change we can build with libkrun main and run a vm with
vment-helper example script or minikube krunkit driver.

I think this should be fixed in libkrun. This seems to be a regression
introduced by https://github.com/containers/libkrun/pull/367. But the
fix in krunki is easy and safe.

Signed-off-by: Nir Soffer <nirsof@gmail.com>