context: check if restful URI scheme is None
Before starting the status listener thread, we not only need to check if
the RestfulUri variant != None, but we also need to verify the user
specified a restful URI scheme at all.

This prevents a panic when we call `status_listener()` due to hitting
an `unreachable!()` code path. Functionally, the panic didn't affect users;
it's difficult to catch since the VM starts immediately after the
message is printed to stdout.

Signed-off-by: Jake Correnti <jakecorrenti+github@proton.me>